Moved the fonts folder to Assets so it can be loaded via Resources.load(). Added textures to some models. Added Diamond spinny thing. Added Stun. Changed to floats.

+++ Assets/Scripts/Tower Scripts/Stats/BaseStat.cs
@@ -2,7 +2,7 @@
 
 public class BaseStat {
 	// Stat Properties
-	private int _baseValue;			// Base value of this stat.
+	private float _baseValue;			// Base value of this stat.
 
 
 	public BaseStat (int val) {
@@ -13,7 +13,7 @@ public class BaseStat {
 		_baseValue = 0;
 	}
 
-	public int BaseValue {
+	public float BaseValue {
 		get{ return _baseValue; }
 		set{ _baseValue = value; }
 	}
